Build and run using Docker
Motivation:
Make it easier to use modex on any platform by building and running it
in a Linux container.

Modifications:
Added a Dockerfile based on debian which can be built and run using the
instructions added to the README. Tested on macOS.

The instructions for building mikmod were not quite up to date, at least I
needed to install autotools and makeinfo to get it working. I used help from
these sites for writing the Dockerfile:

- https://developer.gnome.org/anjuta-build-tutorial/stable/create-autotools.html.en
- https://stackoverflow.com/questions/22603163/automake-error-ltmain-sh-not-found
- https://stackoverflow.com/questions/338317/what-is-makeinfo-and-how-do-i-get-it
- https://unix.stackexchange.com/questions/22926/where-do-executables-look-for-shared-objects-at-runtime